requirements.txt
openai==1.12.0
pandas
tkinter
pillow
request
retrying
